pyproject.toml
[project]
name = "expertlens"
version = "0.1.0"
description = "Real-time multimodal AI coaching agent using Gemini Live API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.19.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "PyJWT>=2.8.0",
    "pwdlib[argon2]>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agent"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyright>=1.1.0",
    "ruff>=0.8.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"agent/prompts/**" = ["E501"]  # prompt templates contain long natural-language sentences
